refactor FieldBackedProvider
